+++ src/app/cart/components/cart-list/cart-list.component.ts
@@ -4,6 +4,7 @@ import {CommonModule} from '@angular/common';
 import {CartItemModel} from "../../models/cart-item.model";
 import {CartService} from "../../services/cart.service";
 import {CartItemComponent} from "../cart-item/cart-item.component";
+import {ProductModel} from "../../../products/models/product. model";
 
 @Component({
   selector: 'app-cart-list',
@@ -16,4 +17,16 @@ export class CartListComponent {
   cartService: CartService = inject(CartService)
 
   trackByItems(index: number, item: CartItemModel): string { return item.name }
+
+  onQuantityIncrease(cartItem: ProductModel) {
+    this.cartService.increaseAmount(cartItem);
+  }
+
+  onQuantityDecrease(cartItem: ProductModel) {
+    this.cartService.decreaseAmount(cartItem);
+  }
+
+  onDeleteItem(cartName: string) {
+    this.cartService.deleteItem(cartName)
+  }
 }