+++ src/app/cart/services/cart.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@angular/core';
+import {computed, Injectable, signal, Signal, WritableSignal} from '@angular/core';
 import {CartItemModel} from "../models/cart-item.model";
 import {ProductModel} from "../../products/models/product. model";
 
@@ -7,31 +7,60 @@ import {ProductModel} from "../../products/models/product. model";
 })
 export class CartService {
 
-  cartList: CartItemModel[] = [];
+  cartList: WritableSignal<CartItemModel[]> = signal<CartItemModel[]>([]);
 
-  get totalPrice(): number {
-    return this.cartList
-      .map(item => item.price)
-      .reduce((prev, acc) => prev + acc);
-  }
+  totalPrice: Signal<number> = computed(() =>
+    this.cartList()
+      .map(item => item.price * item.amount)
+      .reduce((prev, acc) => prev + acc))
+
+  totalAmount: Signal<number> = computed(() =>
+    this.cartList()
+      .map(item => item.amount)
+      .reduce((prev, acc) => prev + acc)
+  )
 
   get cartItems(): CartItemModel[] {
-    return this.cartList;
+    return this.cartList();
   }
 
   addItemToCart(item: ProductModel) {
-    const existingProduct = this.cartList.find(el => item.name === el.name)
+    const existingProduct = this.cartList().findIndex(el => item.name === el.name)
 
-    if (existingProduct) {
-      existingProduct.amount++;
-      existingProduct.price = existingProduct.price + item.price
+    if (existingProduct !== -1) {
+      this.increaseAmount(item)
     } else {
       const newItem = {
         name: item.name,
         price: item.price,
         amount: 1
       }
-      this.cartList = [...this.cartList, newItem]
+      this.cartList.update(value => [...value, newItem])
     }
   }
+
+  increaseAmount(cartItem: ProductModel) {
+    this.cartList.update(value =>
+      value.map(el => (el.name === cartItem.name)
+        ? { ...el, amount: el.amount + 1 }
+        : el
+      )
+    );
+  }
+
+  decreaseAmount(cartItem: ProductModel) {
+      this.cartList.update(value =>
+        value.map(el => (el.name === cartItem.name)
+          ? { ...el, amount: el.amount - 1 }
+          : el
+        )
+      );
+  }
+
+  deleteItem(cartName: string) {
+    this.cartList.update(value =>
+      value.filter(el => el.name !== cartName)
+    )
+  }
+
 }