+++ src/app/products/components/product/product.component.ts
@@ -1,15 +1,15 @@
-import {Component, inject, Input} from '@angular/core';
+import {Component, EventEmitter, Input, Output} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {MatButtonModule} from "@angular/material/button";
 import {MatCardModule} from "@angular/material/card";
 
 import {ProductModel} from "../../models/product. model";
-import {CartService} from "../../../cart/services/cart.service";
+import {HighlightDirective} from "../../directives/highlight.directive";
 
 @Component({
   selector: 'app-product',
   standalone: true,
-  imports: [CommonModule, MatButtonModule, MatCardModule],
+  imports: [CommonModule, MatButtonModule, MatCardModule, HighlightDirective],
   templateUrl: './product.component.html',
   styleUrls: ['./product.component.scss']
 })
@@ -17,11 +17,5 @@ export class ProductComponent {
   @Input({
     required: true
   }) product: ProductModel;
-
-  cartService: CartService = inject(CartService);
-
-  onAddToCard() {
-    console.log(`${this.product.name} was added to your card!`)
-    this.cartService.addItemToCart(this.product)
-  }
+  @Output() onAddToCart = new EventEmitter();
 }